requirements.txt
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astropy==7.1.0
astropy-iers-data==0.2025.9.29.0.35.48
astroquery==0.4.11
attrs==25.3.0
azure-core==1.35.1
azure-storage-blob==12.26.0
beautifulsoup4==4.14.2
bokeh==3.8.0
botocore==1.40.18
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
fastapi==0.118.0
fbpca==1.0
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
html5lib==1.1
idna==3.10
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
keyring==25.6.0
kiwisolver==1.4.9
lightkurve==2.5.1
MarkupSafe==3.0.3
matplotlib==3.10.6
memoization==0.4.0
more-itertools==10.8.0
multidict==6.6.4
narwhals==2.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
propcache==0.4.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pyerfa==2.0.1.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyvo==1.7
PyYAML==6.0.3
requests==2.32.5
s3fs==2025.9.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.5.0
uvicorn==0.37.0
webencodings==0.5.1
wrapt==1.17.3
xyzservices==2025.4.0
yarl==1.20.1
